+++ app/assets/javascripts/discourse/app/components/group-member-dropdown.js
@@ -11,56 +11,7 @@ export default DropdownSelectBoxComponent.extend({
     showFullTitle: false,
   },
 
-  contentBulk() {
-    const items = [];
-
-    items.push({
-      id: "removeMembers",
-      name: I18n.t("groups.members.remove_members"),
-      description: I18n.t("groups.members.remove_members_description"),
-      icon: "user-times",
-    });
-
-    if (this.bulkSelection.some((m) => !m.owner)) {
-      items.push({
-        id: "makeOwners",
-        name: I18n.t("groups.members.make_owners"),
-        description: I18n.t("groups.members.make_owners_description"),
-        icon: "shield-alt",
-      });
-    }
-
-    if (this.bulkSelection.some((m) => m.owner)) {
-      items.push({
-        id: "removeOwners",
-        name: I18n.t("groups.members.remove_owners"),
-        description: I18n.t("groups.members.remove_owners_description"),
-        icon: "shield-alt",
-      });
-    }
-
-    if (this.bulkSelection.some((m) => !m.primary)) {
-      items.push({
-        id: "setPrimary",
-        name: I18n.t("groups.members.make_all_primary"),
-        description: I18n.t("groups.members.make_all_primary_description"),
-        icon: "id-card",
-      });
-    }
-
-    if (this.bulkSelection.some((m) => m.primary)) {
-      items.push({
-        id: "unsetPrimary",
-        name: I18n.t("groups.members.remove_all_primary"),
-        description: I18n.t("groups.members.remove_all_primary_description"),
-        icon: "id-card",
-      });
-    }
-
-    return items;
-  },
-
-  contentSingle() {
+  content: computed("member.owner", "member.primary", function () {
     const items = [
       {
         id: "removeMember",
@@ -117,16 +68,5 @@ export default DropdownSelectBoxComponent.extend({
     }
 
     return items;
-  },
-
-  content: computed(
-    "bulkSelection.[]",
-    "member.owner",
-    "member.primary",
-    function () {
-      return this.bulkSelection !== undefined
-        ? this.contentBulk()
-        : this.contentSingle();
-    }
-  ),
+  }),
 });
