+++ app/assets/javascripts/discourse/app/components/table-header-toggle.js
@@ -5,7 +5,7 @@ import { htmlSafe } from "@ember/template";
 export default Component.extend({
   tagName: "th",
   classNames: ["sortable"],
-  attributeBindings: ["title"],
+  attributeBindings: ["title", "colspan"],
   labelKey: null,
   chevronIcon: null,
   columnIcon: null,
